✨ Provision UPS based on CR
JIRA: https://issues.jboss.org/browse/AEROGEAR-9009

This doesn't yet create the Ingress or Route, since it's not 100%
clear what we need to do around that, and the interaction with the
oauth-proxy container in the UPS Pod.

Kubernetes doesn't appear to provide an OAuth server ootb like
OpenShift does, so the OpenShift oauth-proxy that we use may not work
there as easily.

Also, when setting the OAuthRedirectReference in the ServiceAccount,
the OpenShift docs currently claim that Route is the only supported
Kind at the moment.

There's also the more obvious differences between Route and Ingress:
we're used to using Route, where we just give it a name and the
fully-qualified URL gets provided to us automatically, and configured
in the Router. For k8s Ingress, we need to provide the host directly
in the Ingress resource, so for the operator to do that, it would need
to know what that should be: so it would probably need to be provided
by the person provisioning, in the UnifiedPushServer CR.