Add ServiceMonitor definition for Prometheus
JIRA: https://issues.jboss.org/browse/AEROGEAR-9011

Running `make cluster/prepare` will now label the namespace with
`monitoring_key=middleware`, which the prometheus-operator provisioned
by the application-monitoring-operator[1] will look out for.

The prometheus-operator can read the ServiceMonitor CR, and instruct
the Prometheus instance to start scraping metrics as appropriate.

This change doesn't create any GrafanaDashboard or PrometheusRule
CRs. Since we're just exposing the default metrics from an operator,
it should be up to the SRE team to decide what they want to be able to
see and be alerted by, and they would likely want this to be the same
for all operators, rather than what _we_ decide.

[1] https://github.com/integr8ly/application-monitoring-operator