Add missing internal Service to provisioned UPS
JIRA: https://issues.jboss.org/browse/AEROGEAR-9022

This is needed for the operator to be able to interact with UPS within
the cluster, without needing to go through the oauth-proxy.

This also adds a new label to it with a key of "internal" and a value
of "unifiedpush", this is for filtering the correct one more easily.