Only watch relevant namespaces for CRs
Before this change, we'd try to reconcile a CR in any namespace. This
changes that so that the controller doesn't even bother calling the
Reconcile method if the CR isn't in one of the namespaces that we
expect for that kind.

Since running the operator locally with operator-sdk up local doesn't
seem to correctly react to CRs in namespaces other than one for the
current context, this change also adds that to the list of
APP_NAMESPACES for the `make code/run` target.

Co-authored-by: Laura Fitzgerald <lfitzger@redhat.com>