Use annotations for storing Ids; reconcile secrets
We previously relied on Ids (from UPS) stored in the Status block of
the CRs, but that led to problems when trying to restore from
backups.

This change doesn't remove the Ids from the Status block, since we
know the MDC relies on them being there for the moment. We should
consider updating that to get it from the annotation, and then no
longer present it in the Status block.

I've also updated some of the reconciliation logic to update the
secrets in the Status block, if someone has performed a "Renew Secret"
through the UPS Admin UI.