🐛 Use Deployment readyReplicas to check readiness
JIRA: https://issues.jboss.org/browse/AEROGEAR-9958

The old version checked a condition, which seemed to get set to true
early on (possibly when it successfully created its replicaset), and
deleting all pods didn't unset it.

This change now considers a Deployment "ready" if it has at least one
replica/pod ready. This should be the point at which these pods are
load-balanced by the Service, so even if not _all_ replicas are ready,
there's enough that the application is "ready" and open for business.

Also, I had previously set a property in the reconcile struct to be
able to access the ready status, but that struct is actually just
initialized once, and then re-used for all invocations of its
Reconcile method -- this doesn't help with how I was using it, I
always expect it to be initially true coming into the function for the
boolean logic to work.